version 2: look for only shared lib first
The first version didn't work at all under macOS.

This is another take on the problem: this searches for only a shared
library first, and only if that fails opens up the search for a static
library.  The shared library, if found, is set in
PYTHON_LIBRARY_SHARED; the embedded tests now check that to see whether
a shared python lib was actually found, and if not, skips the embedded
tests.