Don't attempt to constructor an impossible shared_ptr
Fixes #1178

It's possible for a non-destructible base class to be declared; such a
holder *can* be obtained via `std::enable_shared_from_this`, but a
shared_ptr cannot be constructed from a returned pointer.

This commit puts the holder constructor behind a `std::is_destructible`
check, giving a runtime failure if we are ever given such a pointer
without a valid `shared_from_this()` shared pointer.