Enable `enable_shared_from_this` support for casts
Currently we support `enable_shared_from_this` only if we end up taking
over an existing reference or pointer; this extends the support by
looking for `std::enable_shared_shared_from_this` inheritance for any
type of cast except for an explicit `return_value_policy::copy`.

This lets pybind avoid a copy when doing something like:

    m.def("f", []() {
        auto b = std::make_shared<B>();
        foo(*b);
    }

where `foo` is some function that takes the `B` by lvalue reference,
but ends up requiring casting it to Python.  (In the motivating example
for this PR the specific case was a trampoline overload).

Cc: @florianwechsung for input.