Don't add duplicate patients
This fixes #1251 (patient vector grows without bounds) for the 2.2.2
branch by checking that the vector doesn't already have the given
patient.

This is a little less elegant than the same fix for `master` in #1253 (which
changes the patients `vector` to an `unordered_set`), but that requires
an internals layout change, which this approach avoids.