Detect virtual bases as non-simple types
We currently automatically detect multiple inheritance (when registered
with pybind) to enable non-simple casting, but simple casting is also
invalid when there is single-base, virtual inheritance (#1256).  (This
happens to work under gcc/clang, but makes MSVC segfault when upcasting).

This commit adds detection for a specified base class being a virtual
base, and if so, turns on the `multiple_inheritance` flag to mark the
class and base as non-simple, and adds tests (which segfault under MSVC
without this fix).

(Earlier versions can work around the issue by explicitly passing a
`py::multiple_inheritance{}` option to the `py::class_<...>` constructor).

Co-authored-by: Jason Rhinelander <jason@imaginary.ca>