Correctly report Python exceptions set but not thrown during import.
Consider the following example:

    #include <pybind11/pybind11.h>
    PYBIND11_MODULE(python_example, m) {
        PyErr_SetString(PyExc_ImportError, "oops");
    }

Currently, importing this module will result in

    SystemError: initialization of python_example raised unreported exception

because the Python exception is not converted to a C++ one
(py::error_already_set).  Fix that by checking the error flag, so that
one instead gets

    ImportError: oops

Real use cases include numpy's `import_array()` or pycairo's
`import_cairo()`, which can set the exception flag (and return -1) on
failure.  Currently the correct way to handle them with pycairo is

    if (import_foo() < 0) { m.ptr() = nullptr; return; }

and this patch removes the need for `m.ptr() = nullptr;`.