Allow a Python exception to be raised without throwing.
While probably not a big deal in general code, this might result in
significant speedups in critical code paths, such as raising an
IndexError in __getitem__() implementations to stop iterating --
for example conversion of small arrays or math vector classes to
lists (or np.array, for example).

Measuring with Magnum's Python bindings and the timeit
module, the numbers were as follows (Release build, GCC 9, Py3.7):

    Vector3(1.0, 2.0, 3.0)                        0.54144 µs
    list(Vector3(1.0, 2.0, 3.0))                  5.25274 µs
    Vector3()[0] # doesn't throw                  0.97913 µs
    Vector3()[3] # throws                         5.78696 µs
    raise IndexError() # throws                   0.15339 µs

After this patch and a change where Vector3's __getitem__ calls
PyErr_SetString() and returns an empty py::object instead of throwing
py::index_error, the numbers went down considerably:

    Vector3(1.0, 2.0, 3.0)                        0.55200 µs
    list(Vector3(1.0, 2.0, 3.0))                  1.60032 µs
    Vector3()[0] # doesn't throw                  0.68562 µs
    Vector3()[3] # throws                         0.84481 µs
    raise IndexError() # throws                   0.17830 µs

To be extra sure about the impact, I also tried PyErr_SetString()
together with throwing py::error_already_set as that's also less code
being executed, but that led only to comparably minor improvement
(~4 µs down from 5.8).

Important thing to note is that for conversion to np.array and friends,
implementing a buffer protocol is far more performant solution (under a
microsecond compared to around 6 µs when relying on this patch and
almost 20 µs when throwing py::index_error). So this change is mainly
aimed at list conversions and general iteration.