Catch exceptions thrown by repr() when called to print a "incompatible arguments" error.
The dispatcher currently calls repr() if no overload matches, but repr() may throw causing the process to terminate with an uncaught exception. Handle that case more gracefully by falling back to a string that says that repr threw. With more work we could describe the error but that seems unnecessary, given the caller can just call repr() on the argument themselves if they want.