Improve the documentation of partitionning/extending extension modules.
Why? I have been looking for this for hours, and even though I had read
the section, I had not understood it, as "everything 'just' works" was
not explicit enough for me.

I have seen a similar question in the Gitter chat, with a great example.

This proves that I was not alone, and I feel this is an advanced topic
that could benefit from being a little more expicit.

What? Thus, this explicits the problem with referecing variables
defined in another scope, and this also gives a full example
with the different imports.

Hopefuly, the will save some time to some people.