Update make_object_base_type implementation to handle exceptions
When the exception flag is passed, we need to inherit from
PyExc_Exception instead of from PyBaseObject_Type. Since PyExc_Exception
isn't recognized as a PyTypeObject, we have to cast it.

Since exceptions support dynamic attributes, we have to make sure that
GC is turned on for classes which are following that class path.

Update make_new_python_type to choose the exception_base type when
is_except has been set.